PEAR bug 21082: Use BSD license
This fixes a discrepancy in the licensing of the Mail package. In
PEAR bug #15620 the license was changed from PHP 2.02 to the
BSD-2-clause, but somehow the license in sendmail.php wasn't changed.
A new bug was opened for this: PEAR bug #21082. Permission for the
license change was given by the authors in both bug #15620 and #21082.

See also: https://pear.php.net/bugs/bug.php?id=15620
See also: https://pear.php.net/bugs/bug.php?id=21082